+++ index.html
@@ -11,7 +11,7 @@
     <link href="https://fonts.googleapis.com/css2?family=Fira+Code&family=Noto+Sans+SC:wght@100;300;400&display=swap" rel="stylesheet">
     <!-- Functionality -->
     <script src="static/js/books.js"></script>
-    <script src="static/js/funcitons.js" type="module"></script>
+    <script src="static/js/functions.js" type="module"></script>
     
     
 </head>
@@ -20,11 +20,11 @@
 
     <!-- Paragraphs to read and write -->
     <div>
-        <div id="div1" class="gray-box"><span></span></div>
-        <div id="div2" class="gray-box"><span></span></div>
+        <div id="div1" class="gray-box"></div>
+        <div id="div2" class="gray-box"></div>
         <div id="div3" class="white-box" onkeypress="checkCorrect()"></div>
-        <div id="div4" class="gray-box"><span></span></div>
-        <div id="div5" class="gray-box"><span></span></div>
+        <div id="div4" class="gray-box"></div>
+        <div id="div5" class="gray-box"></div>
     </div>
 
     <!-- Book progress -->
